+++ plugins/woocommerce-admin/packages/components/CHANGELOG.md
@@ -1,9 +1,13 @@
 # Unreleased
 
--   Change styling of `<ProductImage />`
+-   Change styling of `<ProductImage />`.
 -   Add new `<Accordion>` component.
 -   Remove the `showCount` prop from `<SearchListItem>`. Count will always be displayed if any of those props is not undefined/null: `countLabel` and `item.count`.
 -   Fix alignment of `<SearchListItem>` count bubble in newest versions of `@wordpress/components`.
+-   Support custom attributes in `<AttributeFilter />`.
+-   Add product attributes support to `<Search />`.
+-   Allow single-selection support to `<Search />`.
+-   Improve handling of `multiple` and `inlineTags` in `<SelectControl />`.
 
 # 5.1.2